pyproject.toml
[tool.poetry]
name = "toxicitylimitations"
version = "0.1.0"
description = "code used for the paper about the limitations of toxicity detection systems"
authors = ["TheophileBERTELOOT <thber64@ulaval.ca>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ToxicityLimitations"}]


[tool.poetry.scripts]
main = "ToxicityLimitations.main:cli"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
perspectiveapi = "^1.1.0"
openai = "^1.52.2"
transformers = "^4.46.0"
torch = "^2.5.0"
perspective = "^1.0.3"
datasets = "^3.0.2"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
